+++ summon-core/src/commonMain/kotlin/codes/yousef/summon/modifier/MediaQueryModifiers.kt
@@ -252,7 +252,9 @@ fun Modifier.mediaQuery(query: MediaQuery, builder: Modifier.() -> Modifier): Mo
         "$existingQueries|$queryString{$stylesString}"
     }
 
-    return attribute("data-media-queries", newQueries)
+    return withConditionalStyle(
+        MediaStyleDefinition(query, mediaModifier.styles)
+    ).attribute("data-media-queries", newQueries)
 }
 
 /**